refactor #148: move page container to atoms folder
