+++ app/leaderboard/page.tsx
@@ -8,7 +8,7 @@ import PageHeader from '@/components/pageHeader'
 import { statsColumns } from '@/constants'
 import { headers } from 'next/headers'
 import ColumnCards from '@/components/columnCards'
-import { PageContainer } from '@/components/pageContainer'
+import { PageContainer } from '@/components/atoms/pageContainer'
 import Link from '@/components/link'
 import ColumnCardsWithControls from '@/components/columnCardsWithControls'
 import Info from '@/components/atoms/info'
