+++ components/columnCardsWithControls/index.tsx
@@ -2,7 +2,7 @@
 
 import { BestTime } from '@/interfaces/player'
 import ColumnCards from '../columnCards'
-import { PageContainer } from '../pageContainer'
+import { PageContainer } from '../atoms/pageContainer'
 import Badges from '../atoms/badges'
 import { BadgesContainer } from './styled'
 import { useState, useEffect } from 'react'
