+++ components/scoreboardSelector/index.tsx
@@ -3,7 +3,7 @@
 import { useTheme } from '@/hooks/useTheme'
 import { Clock } from '../icons/clock'
 import { Stats } from '../icons/stats'
-import { PageContainer } from '../pageContainer'
+import { PageContainer } from '../atoms/pageContainer'
 import Toggle from '../atoms/toggle'
 import { useRouter } from 'next/navigation'
